Fix null control_stream.on_recv
When I ran init_ipython_kernel() I was getting an error stating that the
IPython kernel didn't have a `control_stream`, and so
`control_stream.on_recv` was throwing a null pointer.

It seems that when `OurIPythonKernel` is initialized, it's passing along
a deprecated arg (`shell_streams`), and the control_stream is never attached to the
IPython KernelBase. Passing it along explicitly solves the problem.

For reference, the shell_streams arg was deprecated in ipython 6.0
https://github.com/ipython/ipykernel/blob/7ed57cb4a03e719074c373029cf9f2e6dfd2c6de/ipykernel/kernelbase.py#L65

library versions used:
ipykernel-6.0.3
ipython-7.25.0